+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### 🐛 Fixes
 - [6296](https://github.com/vegaprotocol/vega/issues/6296) - Add `assetId` as a field to the `GraphQL UpdateAsset` type
-- [](https://github.com/vegaprotocol/vega/issues/xxxx) -
+- [6273](https://github.com/vegaprotocol/vega/issues/6273) - Deposits return an empty list when no party is provided
 
 
 ## 0.56.0